@@ -1,9 +1,5 @@
 @preconcurrency import Dispatch
-#if !canImport(Darwin) && swift(<5.9)
-@preconcurrency import Foundation
-#else
 import Foundation
-#endif
 import NIOCore
 import NIOConcurrencyHelpers
 
@@ -11,6 +7,11 @@ extension Response {
     struct BodyStream: Sendable {
         let count: Int
         let callback: @Sendable (BodyStreamWriter) -> ()
+    }
+    
+    struct AsyncBodyStream {
+        let count: Int
+        let callback: @Sendable (AsyncBodyStreamWriter) async throws -> ()
     }
 
     /// Represents a `Response`'s body.
@@ -29,6 +30,7 @@ extension Response {
             case staticString(StaticString)
             case string(String)
             case stream(BodyStream)
+            case asyncStream(AsyncBodyStream)
         }
         
         /// An empty `Response.Body`.
@@ -56,6 +58,7 @@ extension Response {
             case .buffer(let buffer): return buffer.readableBytes
             case .none: return 0
             case .stream(let stream): return stream.count
+            case .asyncStream(let stream): return stream.count
             }
         }
         
@@ -69,6 +72,7 @@ extension Response {
             case .string(let string): return Data(string.utf8)
             case .none: return nil
             case .stream: return nil
+            case .asyncStream: return nil
             }
         }
         
@@ -89,6 +93,7 @@ extension Response {
                 return buffer
             case .none: return nil
             case .stream: return nil
+            case .asyncStream: return nil
             }
         }
 
@@ -99,6 +104,13 @@ extension Response {
                 stream.callback(collector)
                 return collector.promise.futureResult
                     .map { $0 }
+            case .asyncStream(let stream):
+                let collector = ResponseBodyCollector(eventLoop: eventLoop, byteBufferAllocator: self.byteBufferAllocator)
+                return eventLoop.makeFutureWithTask {
+                    try await stream.callback(collector)
+                }.flatMap {
+                    collector.promise.futureResult.map { $0 }
+                }
             default:
                 return eventLoop.makeSucceededFuture(self.buffer)
             }
@@ -114,6 +126,7 @@ extension Response {
             case .staticString(let string): return string.description
             case .string(let string): return string
             case .stream: return "<stream>"
+            case .asyncStream: return "<async stream>"
             }
         }
         
@@ -167,6 +180,54 @@ extension Response {
             self.init(stream: stream, count: -1, byteBufferAllocator: byteBufferAllocator)
         }
         
+        /// Creates a chunked HTTP ``Response`` steam using ``AsyncBodyStreamWriter``.
+        ///
+        /// - Parameters:
+        ///   - asyncStream: The closure that will generate the results. **MUST** call `.end` or `.error` when terminating the stream
+        ///   - count: The amount of bytes that will be written. The `asyncStream` **MUST** produce exactly `count` bytes.
+        ///   - byteBufferAllocator: The allocator that is preferred when writing data to SwiftNIO
+        public init(asyncStream: @escaping @Sendable (AsyncBodyStreamWriter) async throws -> (), count: Int, byteBufferAllocator: ByteBufferAllocator = ByteBufferAllocator()) {
+            self.byteBufferAllocator = byteBufferAllocator
+            self.storage = .asyncStream(.init(count: count, callback: asyncStream))
+        }
+        
+        /// Creates a chunked HTTP ``Response`` steam using ``AsyncBodyStreamWriter``.
+        ///
+        /// - Parameters:
+        ///   - asyncStream: The closure that will generate the results. **MUST** call `.end` or `.error` when terminating the stream
+        ///   - byteBufferAllocator: The allocator that is preferred when writing data to SwiftNIO
+        public init(asyncStream: @escaping @Sendable (AsyncBodyStreamWriter) async throws -> (), byteBufferAllocator: ByteBufferAllocator = ByteBufferAllocator()) {
+            self.init(asyncStream: asyncStream, count: -1, byteBufferAllocator: byteBufferAllocator)
+        }
+        
+        /// Creates a _managed_ chunked HTTP ``Response`` steam using ``AsyncBodyStreamWriter`` that automtically closes or fails based if the closure throws an error or returns.
+        ///
+        /// - Parameters:
+        ///   - asyncStream: The closure that will generate the results, which **MUST NOT** call `.end` or `.error` when terminating the stream.
+        ///   - count: The amount of bytes that will be written. The `asyncStream` **MUST** produce exactly `count` bytes.
+        ///   - byteBufferAllocator: The allocator that is preferred when writing data to SwiftNIO
+        public init(managedAsyncStream: @escaping @Sendable (AsyncBodyStreamWriter) async throws -> (), count: Int, byteBufferAllocator: ByteBufferAllocator = ByteBufferAllocator()) {
+            self.byteBufferAllocator = byteBufferAllocator
+            self.storage = .asyncStream(.init(count: count) { writer in
+                do {
+                    try await managedAsyncStream(writer)
+                    try await writer.write(.end)
+                } catch {
+                    try await writer.write(.error(error))
+                }
+            })
+        }
+        
+        /// Creates a _managed_ chunked HTTP ``Response`` steam using ``AsyncBodyStreamWriter`` that automtically closes or fails based if the closure throws an error or returns.
+        ///
+        /// - Parameters:
+        ///   - asyncStream: The closure that will generate the results, which **MUST NOT** call `.end` or `.error` when terminating the stream.
+        ///   - count: The amount of bytes that will be written
+        ///   - byteBufferAllocator: The allocator that is preferred when writing data to SwiftNIO
+        public init(managedAsyncStream: @escaping @Sendable (AsyncBodyStreamWriter) async throws -> (), byteBufferAllocator: ByteBufferAllocator = ByteBufferAllocator()) {
+            self.init(managedAsyncStream: managedAsyncStream, count: -1, byteBufferAllocator: byteBufferAllocator)
+        }
+        
         /// `ExpressibleByStringLiteral` conformance.
         public init(stringLiteral value: String) {
             self.byteBufferAllocator = ByteBufferAllocator()
@@ -184,7 +245,7 @@ extension Response {
 // Since all buffer mutation is done on the event loop, we can be unchecked here.
 // This removes the need for a lock and performance hits from that
 // Any changes to this type need to be carefully considered
-private final class ResponseBodyCollector: BodyStreamWriter, @unchecked Sendable {
+private final class ResponseBodyCollector: BodyStreamWriter, AsyncBodyStreamWriter, @unchecked Sendable {
     var buffer: ByteBuffer
     let eventLoop: EventLoop
     let promise: EventLoopPromise<ByteBuffer>
@@ -209,5 +270,12 @@ private final class ResponseBodyCollector: BodyStreamWriter, @unchecked Sendable
         }
         // Fixes an issue where errors in the stream should fail the individual write promise.
         if let promise { future.cascade(to: promise) }
+    }
+    
+    func write(_ result: BodyStreamResult) async throws {
+        let promise = self.eventLoop.makePromise(of: Void.self)
+        
+        self.eventLoop.execute { self.write(result, promise: promise) }
+        try await promise.futureResult.get()
     }
 }
