@@ -7,11 +7,15 @@ struct Entrypoint {
         var env = try Environment.detect()
         try LoggingSystem.bootstrap(from: &env)
 
-        let app = Application(env)
-        defer { app.shutdown() }
-
-        try configure(app)
-        try await app.execute()
+        let app = try await Application.make(env)
+        do {
+            try configure(app)
+            try await app.execute()
+            try await app.asyncShutdown()
+        } catch {
+            try? await app.asyncShutdown()
+            throw error
+        }
     }
 }
 
