@@ -4,10 +4,6 @@ import NIOConcurrencyHelpers
 import NIOCore
 import NIOPosix
 
-#if swift(<5.9)
-    import Backtrace
-#endif
-
 /// Core type representing a Vapor application.
 public final class Application: Sendable {
     public var environment: Environment {
@@ -18,7 +14,7 @@ public final class Application: Sendable {
             self._environment.withLockedValue { $0 = newValue }
         }
     }
-
+    
     public var storage: Storage {
         get {
             self._storage.withLockedValue { $0 }
@@ -27,11 +23,11 @@ public final class Application: Sendable {
             self._storage.withLockedValue { $0 = newValue }
         }
     }
-
+    
     public var didShutdown: Bool {
         self._didShutdown.withLockedValue { $0 }
     }
-
+    
     public var logger: Logger {
         get {
             self._logger.withLockedValue { $0 }
@@ -40,18 +36,18 @@ public final class Application: Sendable {
             self._logger.withLockedValue { $0 = newValue }
         }
     }
-
+    
     public struct Lifecycle: Sendable {
         var handlers: [LifecycleHandler]
         init() {
             self.handlers = []
         }
-
+        
         public mutating func use(_ handler: LifecycleHandler) {
             self.handlers.append(handler)
         }
     }
-
+    
     public var lifecycle: Lifecycle {
         get {
             self._lifecycle.withLockedValue { $0 }
@@ -60,17 +56,17 @@ public final class Application: Sendable {
             self._lifecycle.withLockedValue { $0 = newValue }
         }
     }
-
+    
     public final class Locks: Sendable {
         public let main: NIOLock
         // Is there a type we can use to make this Sendable but reuse the existing lock we already have?
         private let storage: NIOLockedValueBox<[ObjectIdentifier: NIOLock]>
-
+        
         init() {
             self.main = .init()
             self.storage = .init([:])
         }
-
+        
         public func lock<Key>(for key: Key.Type) -> NIOLock
         where Key: LockKey {
             self.main.withLock {
@@ -80,7 +76,7 @@ public final class Application: Sendable {
             }
         }
     }
-
+    
     public var locks: Locks {
         get {
             self._locks.withLockedValue { $0 }
@@ -89,21 +85,21 @@ public final class Application: Sendable {
             self._locks.withLockedValue { $0 = newValue }
         }
     }
-
+    
     public var sync: NIOLock {
         self.locks.main
     }
-
+    
     public enum EventLoopGroupProvider: Sendable {
         case shared(EventLoopGroup)
         @available(*, deprecated, renamed: "singleton", message: "Use '.singleton' for a shared 'EventLoopGroup', for better performance")
         case createNew
-
+        
         public static var singleton: EventLoopGroupProvider {
             .shared(MultiThreadedEventLoopGroup.singleton)
         }
     }
-
+    
     public let eventLoopGroupProvider: EventLoopGroupProvider
     public let eventLoopGroup: EventLoopGroup
     internal let isBooted: NIOLockedValueBox<Bool>
@@ -113,14 +109,19 @@ public final class Application: Sendable {
     private let _logger: NIOLockedValueBox<Logger>
     private let _lifecycle: NIOLockedValueBox<Lifecycle>
     private let _locks: NIOLockedValueBox<Locks>
-
-    public init(
+    
+    @available(*, noasync, message: "This initialiser cannot be used in async contexts, use Application.make(_:_:) instead")
+    public convenience init(
         _ environment: Environment = .development,
         _ eventLoopGroupProvider: EventLoopGroupProvider = .singleton
     ) {
-        #if swift(<5.9)
-            Backtrace.install()
-        #endif
+        self.init(environment, eventLoopGroupProvider, async: false)
+        self.asyncCommands.use(self.servers.command, as: "serve", isDefault: true)
+        DotEnvFile.load(for: environment, on: .shared(self.eventLoopGroup), fileio: self.fileio, logger: self.logger)
+    }
+    
+    // async flag here is just to stop the compiler from complaining about duplicates
+    private init(_ environment: Environment = .development, _ eventLoopGroupProvider: EventLoopGroupProvider = .singleton, async: Bool) {
         self._environment = .init(environment)
         self.eventLoopGroupProvider = eventLoopGroupProvider
         switch eventLoopGroupProvider {
@@ -136,7 +137,7 @@ public final class Application: Sendable {
         self._storage = .init(.init(logger: logger))
         self._lifecycle = .init(.init())
         self.isBooted = .init(false)
-        self.core.initialize()
+        self.core.initialize(asyncEnvironment: async)
         self.caches.initialize()
         self.views.initialize()
         self.passwords.use(.bcrypt)
@@ -148,9 +149,14 @@ public final class Application: Sendable {
         self.servers.use(.http)
         self.clients.initialize()
         self.clients.use(.http)
-        self.commands.use(self.servers.command, as: "serve", isDefault: true)
-        self.commands.use(RoutesCommand(), as: "routes")
-        DotEnvFile.load(for: environment, on: .shared(self.eventLoopGroup), fileio: self.fileio, logger: self.logger)
+        self.asyncCommands.use(RoutesCommand(), as: "routes")
+    }
+    
+    public static func make(_ environment: Environment = .development, _ eventLoopGroupProvider: EventLoopGroupProvider = .singleton) async throws -> Application {
+        let app = Application(environment, eventLoopGroupProvider, async: true)
+        await app.asyncCommands.use(app.servers.asyncCommand, as: "serve", isDefault: true)
+        await DotEnvFile.load(for: app.environment, fileio: app.fileio, logger: app.logger)
+        return app
     }
 
     /// Starts the ``Application`` using the ``start()`` method, then waits for any running tasks to complete.
@@ -206,7 +212,7 @@ public final class Application: Sendable {
     /// If you want to run your ``Application`` indefinitely, or until your code shuts the application down,
     /// use ``execute()`` instead.
     public func startup() async throws {
-        try self.boot()
+        try await self.asyncBoot()
 
         let combinedCommands = AsyncCommands(
             commands: self.asyncCommands.commands.merging(self.commands.commands) { $1 },
@@ -219,6 +225,9 @@ public final class Application: Sendable {
         try await self.console.run(combinedCommands, with: context)
     }
 
+    
+    @available(*, noasync, message: "This can potentially block the thread and should not be called in an async context", renamed: "asyncBoot()")
+    /// Called when the applications starts up, will trigger the lifecycle handlers
     public func boot() throws {
         try self.isBooted.withLockedValue { booted in
             guard !booted else {
@@ -229,7 +238,27 @@ public final class Application: Sendable {
             try self.lifecycle.handlers.forEach { try $0.didBoot(self) }
         }
     }
+    
+    /// Called when the applications starts up, will trigger the lifecycle handlers. The asynchronous version of ``boot()``
+    public func asyncBoot() async throws {
+        /// Skip the boot process if already booted
+        guard !self.isBooted.withLockedValue({
+            var result = true
+            swap(&$0, &result)
+            return result
+        }) else {
+            return
+        }
 
+        for handler in self.lifecycle.handlers {
+            try await handler.willBootAsync(self)
+        }
+        for handler in self.lifecycle.handlers {
+            try await handler.didBootAsync(self)
+        }
+    }
+
+    @available(*, noasync, message: "This can block the thread and should not be called in an async context", renamed: "asyncShutdown()")
     public func shutdown() {
         assert(!self.didShutdown, "Application has already shut down")
         self.logger.debug("Application shutting down")
@@ -249,6 +278,36 @@ public final class Application: Sendable {
             self.logger.trace("Shutting down EventLoopGroup")
             do {
                 try self.eventLoopGroup.syncShutdownGracefully()
+            } catch {
+                self.logger.warning("Shutting down EventLoopGroup failed: \(error)")
+            }
+        }
+
+        self._didShutdown.withLockedValue { $0 = true }
+        self.logger.trace("Application shutdown complete")
+    }
+    
+    public func asyncShutdown() async throws {
+        assert(!self.didShutdown, "Application has already shut down")
+        self.logger.debug("Application shutting down")
+
+        self.logger.trace("Shutting down providers")
+        for handler in self.lifecycle.handlers.reversed()  {
+            await handler.shutdownAsync(self)
+        }
+        self.lifecycle.handlers = []
+
+        self.logger.trace("Clearing Application storage")
+        await self.storage.asyncShutdown()
+        self.storage.clear()
+
+        switch self.eventLoopGroupProvider {
+        case .shared:
+            self.logger.trace("Running on shared EventLoopGroup. Not shutting down EventLoopGroup.")
+        case .createNew:
+            self.logger.trace("Shutting down EventLoopGroup")
+            do {
+                try await self.eventLoopGroup.shutdownGracefully()
             } catch {
                 self.logger.warning("Shutting down EventLoopGroup failed: \(error)")
             }
