@@ -68,17 +68,16 @@ extension DecodingError: AbortError {
     /// See `AbortError.reason`
     public var reason: String {
         switch self {
-        case .dataCorrupted(let ctx):
-            return "Data corrupted at path '\(ctx.codingPath.dotPath)'\(ctx.debugDescriptionAndUnderlyingError)"
-        case .keyNotFound(let key, let ctx):
-            let path = ctx.codingPath + [key]
-            return "Value required for key at path '\(path.dotPath)'\(ctx.debugDescriptionAndUnderlyingError)"
-        case .typeMismatch(let type, let ctx):
-            return "Value at path '\(ctx.codingPath.dotPath)' was not of type '\(type)'\(ctx.debugDescriptionAndUnderlyingError)"
-        case .valueNotFound(let type, let ctx):
-            return "Value of type '\(type)' was not found at path '\(ctx.codingPath.dotPath)'\(ctx.debugDescriptionAndUnderlyingError)"
-        @unknown default: return "Unknown error."
+        case let .dataCorrupted(ctx):       return "Data corrupted \(self.contextReason(ctx))"
+        case let .keyNotFound(key, ctx):    return "No such key '\(key.stringValue)' \(self.contextReason(ctx))"
+        case let .typeMismatch(type, ctx):  return "Value was not of type '\(type)' \(self.contextReason(ctx))"
+        case let .valueNotFound(type, ctx): return "No value found (expected type '\(type)') \(self.contextReason(ctx))"
+        @unknown default:                   return "Unknown error"
         }
+    }
+    
+    private func contextReason(_ context: Context) -> String {
+        "at path '\(context.codingPath.dotPath)'\(context.debugDescriptionAndUnderlyingError)"
     }
 }
 
@@ -92,17 +91,13 @@ private extension DecodingError.Context {
         if self.debugDescription.isEmpty {
             return ""
         } else if self.debugDescription.last == "." {
-            return ". \(String(self.debugDescription.dropLast()))"
+            return ". \(self.debugDescription.dropLast())"
         } else {
             return ". \(self.debugDescription)"
         }
     }
     
     private var underlyingErrorDescription: String {
-        if let underlyingError = self.underlyingError {
-            return ". Underlying error: \(underlyingError)"
-        } else {
-            return ""
-        }
+        self.underlyingError.map { ". Underlying error: \(String(describing: $0))" } ?? ""
     }
 }
