@@ -1,7 +1,6 @@
 import NIOCore
 import NIOHTTP1
 import NIOConcurrencyHelpers
-import Atomics
 
 final class HTTPServerResponseEncoder: ChannelOutboundHandler, RemovableChannelHandler {
     typealias OutboundIn = Response
@@ -70,6 +69,26 @@ final class HTTPServerResponseEncoder: ChannelOutboundHandler, RemovableChannelH
                     count: stream.count == -1 ? nil : stream.count
                 )
                 stream.callback(channelStream)
+            case .asyncStream(let stream):
+                let channelStream = ChannelResponseBodyStream(
+                    context: context,
+                    handler: self,
+                    promise: promise,
+                    count: stream.count == -1 ? nil : stream.count
+                )
+                
+                Task {
+                    do {
+                        try await stream.callback(channelStream)
+                        // We assert in ChannelResponseBodyStream that either .end or .error gets sent, so once we
+                        // get here the promise can be assumed to already be completed. However, just in case, succeed
+                        // it here anyway. This guarantees we never leave the callback without completing the promise
+                        // one way or the other in release builds.
+                        promise?.succeed()
+                    } catch {
+                        promise?.fail(error)
+                    }
+                }
             }
         }
     }
@@ -84,18 +103,19 @@ final class HTTPServerResponseEncoder: ChannelOutboundHandler, RemovableChannelH
     }
 }
 
-private final class ChannelResponseBodyStream: BodyStreamWriter {
+private final class ChannelResponseBodyStream: BodyStreamWriter, AsyncBodyStreamWriter {
     let contextBox: NIOLoopBound<ChannelHandlerContext>
     let handlerBox: NIOLoopBound<HTTPServerResponseEncoder>
     let promise: EventLoopPromise<Void>?
     let count: Int?
-    let currentCount: ManagedAtomic<Int>
-    let isComplete: ManagedAtomic<Bool>
+    let currentCount: NIOLoopBoundBox<Int>
+    let isComplete: NIOLockedValueBox<Bool>
     let eventLoop: EventLoop
 
     enum Error: Swift.Error {
         case tooManyBytes
         case notEnoughBytes
+        case apiMisuse // tried to send a buffer or end indication after already ending or erroring the stream
     }
 
     init(
@@ -104,41 +124,79 @@ private final class ChannelResponseBodyStream: BodyStreamWriter {
         promise: EventLoopPromise<Void>?,
         count: Int?
     ) {
+        context.eventLoop.assertInEventLoop()
+        
         self.contextBox = .init(context, eventLoop: context.eventLoop)
         self.handlerBox = .init(handler, eventLoop: context.eventLoop)
         self.promise = promise
         self.count = count
-        self.currentCount = .init(0)
+        self.currentCount = .init(0, eventLoop: context.eventLoop)
         self.isComplete = .init(false)
         self.eventLoop = context.eventLoop
     }
     
+    func write(_ result: BodyStreamResult) async throws {
+        let promise = self.eventLoop.makePromise(of: Void.self)
+        
+        self.eventLoop.execute { self.write(result, promise: promise) }
+        try await promise.futureResult.get()
+    }
+    
+    /// > Note: `self.promise` is the promise that completes the original write to `HTTPServerResponseEncoder` that
+    /// > triggers the streaming response; it should only be succeeded when the stream ends. The `promise` parameter
+    /// > of this method is specific to the particular invocation and signals that a buffer has finished writing or
+    /// > that the stream has been fully completed, and should always be completed or pending completion by the time
+    /// > this method returns. Both promises should be failed when an error occurs, unless otherwise specifically noted.
     func write(_ result: BodyStreamResult, promise: EventLoopPromise<Void>?) {
+        self.eventLoop.assertInEventLoop() // Only check in debug, just in case...
+
+        func finishStream() {
+            self.isComplete.withLockedValue { $0 = true }
+            self.contextBox.value.fireUserInboundEventTriggered(HTTPServerResponseEncoder.ResponseEndSentEvent())
+            // Don't forward the current promise (if any) to the write completion of the end-response signal, as we
+            // will be notified of errors through other paths and can get spurious I/O errors from this write that
+            // ought to be ignored.
+            self.contextBox.value.writeAndFlush(self.handlerBox.value.wrapOutboundOut(.end(nil)), promise: nil)
+        }
+
+        // See https://github.com/vapor/vapor/issues/2976 for why we do some of these checks.
         switch result {
         case .buffer(let buffer):
-            self.contextBox.value.writeAndFlush(self.handlerBox.value.wrapOutboundOut(.body(.byteBuffer(buffer))), promise: promise)
-            if let count = self.count, self.currentCount.wrappingIncrementThenLoad(by: buffer.readableBytes, ordering: .sequentiallyConsistent) > count {
+            guard !self.isComplete.withLockedValue({ $0 }) else { // Don't try to send data if we already ended
+                return promise?.fail(Error.apiMisuse) ?? () // self.promise is already completed, so fail the local one and bail
+            }
+            if let count = self.count, (self.currentCount.value + buffer.readableBytes) > count {
                 self.promise?.fail(Error.tooManyBytes)
-                promise?.fail(Error.notEnoughBytes)
+                promise?.fail(Error.tooManyBytes)
+            } else {
+                self.currentCount.value += buffer.readableBytes
+                // Cascade the completion of the buffer write to the local promise (if any).
+                self.contextBox.value.writeAndFlush(self.handlerBox.value.wrapOutboundOut(.body(.byteBuffer(buffer))), promise: promise)
             }
         case .end:
-            self.isComplete.store(true, ordering: .sequentiallyConsistent)
-            if let count = self.count, self.currentCount.load(ordering: .sequentiallyConsistent) < count {
-                self.promise?.fail(Error.notEnoughBytes)
-                promise?.fail(Error.notEnoughBytes)
+            if !self.isComplete.withLockedValue({ $0 }) { // Don't send the response end events more than once.
+                finishStream()
+                // check this only after sending the stream end; we want to make send that regardless
+                if let count = self.count, self.currentCount.value < count {
+                    self.promise?.fail(Error.notEnoughBytes)
+                    promise?.fail(Error.notEnoughBytes)
+                } else {
+                    self.promise?.succeed()
+                    promise?.succeed()
+                }
+            } else {
+                promise?.fail(Error.apiMisuse) // If we already ended, fail the local promise with API misuse
             }
-            self.contextBox.value.fireUserInboundEventTriggered(HTTPServerResponseEncoder.ResponseEndSentEvent())
-            self.contextBox.value.writeAndFlush(self.handlerBox.value.wrapOutboundOut(.end(nil)), promise: promise)
-            self.promise?.succeed(())
         case .error(let error):
-            self.isComplete.store(true, ordering: .relaxed)
-            self.contextBox.value.fireUserInboundEventTriggered(HTTPServerResponseEncoder.ResponseEndSentEvent())
-            self.contextBox.value.writeAndFlush(self.handlerBox.value.wrapOutboundOut(.end(nil)), promise: promise)
-            self.promise?.fail(error)
+            if !self.isComplete.withLockedValue({ $0 }) { // Don't send the response end events more than once.
+                finishStream()
+                self.promise?.fail(error)
+            }
+            promise?.fail(error) // We want to fail the local promise regardless. Echo the error back.
         }
     }
 
     deinit {
-        assert(self.isComplete.load(ordering: .sequentiallyConsistent), "Response body stream writer deinitialized before .end or .error was sent.")
+        assert(self.isComplete.withLockedValue { $0 }, "Response body stream writer deinitialized before .end or .error was sent.")
     }
 }
