@@ -1,15 +1,17 @@
 import Foundation
 import NIOCore
+import _NIOFileSystem
 
 /// Serves static files from a public directory.
 ///
 /// `FileMiddleware` will default to `DirectoryConfig`'s working directory with `"/Public"` appended.
-public final class FileMiddleware: Middleware {
+public final class FileMiddleware: AsyncMiddleware {
     /// The public directory. Guaranteed to end with a slash.
     private let publicDirectory: String
     private let defaultFile: String?
     private let directoryAction: DirectoryAction
-    
+    private let advancedETagComparison: Bool
+
     public struct BundleSetupError: Equatable, Error {
         
         /// The description of this error.
@@ -22,6 +24,15 @@ public final class FileMiddleware: Middleware {
         public static let publicDirectoryIsNotAFolder: Self = .init(description: "Cannot find any actual folder for the given Public Directory")
     }
 
+    struct ETagHashes: StorageKey {
+        public typealias Value = [String: FileHash]
+
+        public struct FileHash {
+            let lastModified: Date
+            let digestHex: String
+        }
+    }
+
     /// Creates a new `FileMiddleware`.
     ///
     /// - parameters:
@@ -29,16 +40,18 @@ public final class FileMiddleware: Middleware {
     ///     - defaultFile: The name of the default file to look for and serve if a request hits any public directory. Starting with `/` implies
     ///     an absolute path from the public directory root. If `nil`, no default files are served.
     ///     - directoryAction: Determines the action to take when the request doesn't have a trailing slash but matches a directory.
-    public init(publicDirectory: String, defaultFile: String? = nil, directoryAction: DirectoryAction = .none) {
+    ///     - advancedETagComparison: The method used when ETags are generated. If true, a byte-by-byte hash is created (and cached), otherwise a simple comparison based on the file's last modified date and size.
+    public init(publicDirectory: String, defaultFile: String? = nil, directoryAction: DirectoryAction = .none, advancedETagComparison: Bool = false) {
         self.publicDirectory = publicDirectory.addTrailingSlash()
         self.defaultFile = defaultFile
         self.directoryAction = directoryAction
+        self.advancedETagComparison = advancedETagComparison
     }
     
-    public func respond(to request: Request, chainingTo next: Responder) -> EventLoopFuture<Response> {
+    public func respond(to request: Request, chainingTo next: any AsyncResponder) async throws -> Response {
         // make a copy of the percent-decoded path
         guard var path = request.url.path.removingPercentEncoding else {
-            return request.eventLoop.makeFailedFuture(Abort(.badRequest))
+            throw Abort(.badRequest)
         }
 
         // path must be relative.
@@ -46,52 +59,44 @@ public final class FileMiddleware: Middleware {
 
         // protect against relative paths
         guard !path.contains("../") else {
-            return request.eventLoop.makeFailedFuture(Abort(.forbidden))
+            throw Abort(.forbidden)
         }
 
         // create absolute path
         var absPath = self.publicDirectory + path
-
-        // check if path exists and whether it is a directory
-        var isDir: ObjCBool = false
-        guard FileManager.default.fileExists(atPath: absPath, isDirectory: &isDir) else {
-            return next.respond(to: request)
-        }
         
-        if isDir.boolValue {
-            guard absPath.hasSuffix("/") else {
-                switch directoryAction.kind {
-                case .redirect:                    
-                    var redirectUrl = request.url
-                    redirectUrl.path += "/"
-                    return request.eventLoop.future(
-                        request.redirect(to: redirectUrl.string, redirectType: .permanent)
-                    )
-                case .none:
-                    return next.respond(to: request)
+        if let fileInfo = try await FileSystem.shared.info(forFileAt: .init(absPath)) {
+            // path exists, check for directory or file
+            if fileInfo.type == .directory {
+                // directory exists, see if we can return a file
+                if absPath.hasSuffix("/") {
+                    // If a directory, check for the default file
+                    if let defaultFile = defaultFile {
+                        if defaultFile.isAbsolute() {
+                            absPath = self.publicDirectory + defaultFile.removeLeadingSlashes()
+                        } else {
+                            absPath = absPath + defaultFile
+                        }
+                        
+                        if try await FileSystem.shared.info(forFileAt: .init(absPath)) != nil {
+                            // If the default file exists, stream it
+                            return try await request.fileio.asyncStreamFile(at: absPath, advancedETagComparison: advancedETagComparison)
+                        }
+                    }
+                } else {
+                    if directoryAction.kind == .redirect {
+                        var redirectUrl = request.url
+                        redirectUrl.path += "/"
+                        return request.redirect(to: redirectUrl.string, redirectType: .permanent)
+                    }
                 }
-            }
-            
-            // If a directory, check for the default file
-            guard let defaultFile = defaultFile else {
-                return next.respond(to: request)
-            }
-            
-            if defaultFile.isAbsolute() {
-                absPath = self.publicDirectory + defaultFile.removeLeadingSlashes()
             } else {
-                absPath = absPath + defaultFile
-            }
-            
-            // If the default file doesn't exist, pass on request
-            guard FileManager.default.fileExists(atPath: absPath) else {
-                return next.respond(to: request)
+                // file exists, stream it
+                return try await request.fileio.asyncStreamFile(at: absPath, advancedETagComparison: advancedETagComparison)
             }
         }
         
-        // stream the file
-        let res = request.fileio.streamFile(at: absPath)
-        return request.eventLoop.makeSucceededFuture(res)
+        return try await next.respond(to: request)
     }
 
     /// Creates a new `FileMiddleware` for a server contained in an Xcode Project.
