@@ -4,7 +4,7 @@ public protocol RoutesBuilder {
     func add(_ route: Route)
 }
 
-extension UUID: LosslessStringConvertible {
+extension Foundation.UUID: Swift.LosslessStringConvertible {
     public init?(_ description: String) {
         self.init(uuidString: description)
     }
