@@ -1,5 +1,7 @@
-#if os(Linux)
+#if canImport(Glibc)
 import Glibc
+#elseif canImport(Musl)
+import Musl
 #else
 import Darwin.C
 #endif
