@@ -1,5 +1,7 @@
-#if os(Linux)
+#if canImport(Glibc)
 import Glibc
+#elseif canImport(Musl)
+import Musl
 #else
 import Darwin
 #endif
@@ -10,8 +12,7 @@ import NIOPosix
 /// Reads dotenv (`.env`) files and loads them into the current process.
 ///
 ///     let fileio: NonBlockingFileIO
-///     let elg: EventLoopGroup
-///     let file = try DotEnvFile.read(path: ".env", fileio: fileio, on: elg.next()).wait()
+///     let file = try await DotEnvFile.read(path: ".env", fileio: fileio)
 ///     for line in file.lines {
 ///         print("\(line.key)=\(line.value)")
 ///     }
@@ -42,6 +43,7 @@ public struct DotEnvFile: Sendable {
     ///     - eventLoopGroupProvider: Either provides an EventLoopGroup or tells the function to create a new one.
     ///     - fileio: NonBlockingFileIO that is used to read the .env file(s).
     ///     - logger: Optionally provide an existing logger.
+    @available(*, noasync, message: "Use an async version of load instead")
     public static func load(
         for environment: Environment = .development,
         on eventLoopGroupProvider: Application.EventLoopGroupProvider = .singleton,
@@ -89,6 +91,7 @@ public struct DotEnvFile: Sendable {
     ///     - eventLoopGroupProvider: Either provides an EventLoopGroup or tells the function to create a new one.
     ///     - fileio: NonBlockingFileIO that is used to read the .env file(s).
     ///     - logger: Optionally provide an existing logger.
+    @available(*, noasync, message: "Use an async version of load instead")
     public static func load(
         path: String,
         on eventLoopGroupProvider: Application.EventLoopGroupProvider = .singleton,
@@ -219,6 +222,101 @@ public struct DotEnvFile: Sendable {
         for line in self.lines {
             setenv(line.key, line.value, overwrite ? 1 : 0)
         }
+    }
+    
+    // MARK: - Concurrency
+    /// Reads a dotenv file from the supplied path.
+    ///
+    ///     let fileio: NonBlockingFileIO
+    ///     let file = try await DotEnvFile.read(path: ".env", fileio: fileio)
+    ///     for line in file.lines {
+    ///         print("\(line.key)=\(line.value)")
+    ///     }
+    ///     file.load(overwrite: true) // loads all lines into the process
+    ///     print(Environment.process.FOO) // BAR
+    ///
+    /// Use `DotEnvFile.load` to read and load with one method.
+    ///
+    /// - parameters:
+    ///     - path: Absolute or relative path of the dotenv file.
+    ///     - fileio: File loader.
+    public static func read(
+        path: String,
+        fileio: NonBlockingFileIO
+    ) async throws -> DotEnvFile {
+        try await fileio.withFileRegion(path: path) { fileRegion in
+            let buffer = try await fileio.read(fileRegion: fileRegion, allocator: .init())
+            var parser = Parser(source: buffer)
+            return DotEnvFile(lines: parser.parse())
+        }
+    }
+    
+    /// Reads a dotenv file from the supplied path and loads it into the process.
+    ///
+    ///     let fileio: NonBlockingFileIO
+    ///     try await DotEnvFile.load(path: ".env", fileio: fileio)
+    ///     print(Environment.process.FOO) // BAR
+    ///
+    /// Use `DotEnvFile.read` to read the file without loading it.
+    ///
+    /// - parameters:
+    ///     - path: Absolute or relative path of the dotenv file.
+    ///     - fileio: File loader.
+    ///     - overwrite: If `true`, values already existing in the process' env
+    ///                  will be overwritten. Defaults to `false`.
+    public static func load(
+        path: String,
+        fileio: NonBlockingFileIO,
+        overwrite: Bool = false
+    ) async throws {
+        let file = try await self.read(path: path, fileio: fileio)
+        file.load(overwrite: overwrite)
+    }
+    
+    /// Reads the dotenv files relevant to the environment and loads them into the process.
+    ///
+    ///     let path: String
+    ///     let fileio: NonBlockingFileIO
+    ///     let logger: Logger
+    ///     try DotEnvFile.load(path: path, fileio: filio, logger: logger)
+    ///     print(Environment.process.FOO) // BAR
+    ///
+    /// - parameters:
+    ///     - path: Absolute or relative path of the dotenv file.
+    ///     - fileio: NonBlockingFileIO that is used to read the .env file(s).
+    ///     - logger: Optionally provide an existing logger.
+    public static func load(
+        path: String,
+        fileio: NonBlockingFileIO,
+        logger: Logger = Logger(label: "dot-env-loggger")
+    ) async {
+        do {
+            try await load(path: path, fileio: fileio, overwrite: false)
+        } catch {
+            logger.debug("Could not load \(path) file: \(error)")
+        }
+    }
+    
+    /// Reads the dotenv files relevant to the environment and loads them into the process.
+    ///
+    ///     let environment: Environment
+    ///     let fileio: NonBlockingFileIO
+    ///     let logger: Logger
+    ///     try await DotEnvFile.load(for: .development, fileio: fileio, logger: logger)
+    ///     print(Environment.process.FOO) // BAR
+    ///
+    /// - parameters:
+    ///     - environment: current environment, selects which .env file to use.
+    ///     - fileio: NonBlockingFileIO that is used to read the .env file(s).
+    ///     - logger: Optionally provide an existing logger.
+    public static func load(
+        for environment: Environment = .development,
+        fileio: NonBlockingFileIO,
+        logger: Logger = Logger(label: "dot-env-loggger")
+    ) async {
+        // Load specific .env first since values are not overridden.
+        await DotEnvFile.load(path: ".env.\(environment.name)", fileio: fileio, logger: logger)
+        await DotEnvFile.load(path: ".env", fileio: fileio, logger: logger)
     }
 }
 
