@@ -4,10 +4,17 @@ import Vapor
 import NIOCore
 
 final class AsyncCacheTests: XCTestCase {
+    var app: Application!
+    
+    override func setUp() async throws {
+        app = try await Application.make(.testing)
+    }
+    
+    override func tearDown() async throws {
+        try await app.asyncShutdown()
+    }
+    
     func testInMemoryCache() async throws {
-        let app = Application(.testing)
-        defer { app.shutdown() }
-
         let value1 = try await app.cache.get("foo", as: String.self)
         XCTAssertNil(value1)
         try await app.cache.set("foo", to: "bar")
@@ -33,8 +40,6 @@ final class AsyncCacheTests: XCTestCase {
     }
 
     func testCustomCache() async throws {
-        let app = Application(.testing)
-        defer { app.shutdown() }
         app.caches.use(.foo)
         try await app.cache.set("1", to: "2")
         let value = try await app.cache.get("foo", as: String.self)
