@@ -21,11 +21,11 @@ final class AsyncRequestTests: XCTestCase {
     var app: Application!
     
     override func setUp() async throws {
-        app = Application(.testing)
+        app = try await Application.make(.testing)
     }
     
     override func tearDown() async throws {
-        app.shutdown()
+        try await app.asyncShutdown()
     }
     
     func testStreamingRequest() async throws {
@@ -120,13 +120,13 @@ final class AsyncRequestTests: XCTestCase {
                     XCTAssertTrue(serverSawRequest.compareExchange(expected: false, desired: true, ordering: .relaxed).exchanged)
                     var bodyIterator = req.body.makeAsyncIterator()
                     let firstChunk = try await bodyIterator.next() // read only first chunk
-                    numberOfTimesTheServerGotOfferedBytes.wrappingIncrement(ordering: .relaxed)
-                    bytesTheServerSaw.wrappingIncrement(by: firstChunk?.readableBytes ?? 0, ordering: .relaxed)
+                    numberOfTimesTheServerGotOfferedBytes.wrappingIncrement(ordering: .sequentiallyConsistent)
+                    bytesTheServerSaw.wrappingIncrement(by: firstChunk?.readableBytes ?? 0, ordering: .sequentiallyConsistent)
                     defer {
                         _ = bodyIterator // make sure to not prematurely cancelling the sequence
                     }
                     try await Task.sleep(nanoseconds: 10_000_000_000) // wait "forever"
-                    serverSawEnd.store(true, ordering: .relaxed)
+                    serverSawEnd.store(true, ordering: .sequentiallyConsistent)
                     return Response(status: .ok)
                 }
             }
@@ -164,7 +164,7 @@ final class AsyncRequestTests: XCTestCase {
             }
             
             func didSendRequestPart(task: HTTPClient.Task<Response>, _ part: IOData) {
-                self.bytesTheClientSent.wrappingIncrement(by: part.readableBytes, ordering: .relaxed)
+                self.bytesTheClientSent.wrappingIncrement(by: part.readableBytes, ordering: .sequentiallyConsistent)
             }
         }
         
@@ -185,12 +185,12 @@ final class AsyncRequestTests: XCTestCase {
             }
         }
         
-        XCTAssertEqual(1, numberOfTimesTheServerGotOfferedBytes.load(ordering: .relaxed))
-        XCTAssertGreaterThan(tenMB.readableBytes, bytesTheServerSaw.load(ordering: .relaxed))
-        XCTAssertGreaterThan(tenMB.readableBytes, bytesTheClientSent.load(ordering: .relaxed))
-        XCTAssertEqual(0, bytesTheClientSent.load(ordering: .relaxed)) // We'd only see this if we sent the full 10 MB.
-        XCTAssertFalse(serverSawEnd.load(ordering: .relaxed))
-        XCTAssertTrue(serverSawRequest.load(ordering: .relaxed))
+        XCTAssertEqual(1, numberOfTimesTheServerGotOfferedBytes.load(ordering: .sequentiallyConsistent))
+        XCTAssertGreaterThanOrEqual(tenMB.readableBytes, bytesTheServerSaw.load(ordering: .sequentiallyConsistent))
+        XCTAssertGreaterThanOrEqual(tenMB.readableBytes, bytesTheClientSent.load(ordering: .sequentiallyConsistent))
+        XCTAssertEqual(0, bytesTheClientSent.load(ordering: .sequentiallyConsistent)) // We'd only see this if we sent the full 10 MB.
+        XCTAssertFalse(serverSawEnd.load(ordering: .sequentiallyConsistent))
+        XCTAssertTrue(serverSawRequest.load(ordering: .sequentiallyConsistent))
         
         requestHandlerTask.withLockedValue { $0?.cancel() }
         try await httpClient.shutdown()
