@@ -4,7 +4,17 @@ import Vapor
 import NIOHTTP1
 
 final class AsyncRouteTests: XCTestCase {
-    func testEnumResponse() throws {
+    var app: Application!
+    
+    override func setUp() async throws {
+        app = try await Application.make(.testing)
+    }
+    
+    override func tearDown() async throws {
+        try await app.asyncShutdown()
+    }
+    
+    func testEnumResponse() async throws {
         enum IntOrString: AsyncResponseEncodable {
             case int(Int)
             case string(String)
@@ -19,9 +29,6 @@ final class AsyncRouteTests: XCTestCase {
             }
         }
 
-        let app = Application(.testing)
-        defer { app.shutdown() }
-
         app.routes.get("foo") { req -> IntOrString in
             if try req.query.get(String.self, at: "number") == "true" {
                 return .int(42)
@@ -30,22 +37,19 @@ final class AsyncRouteTests: XCTestCase {
             }
         }
 
-        try app.testable().test(.GET, "/foo?number=true") { res in
+        try await app.testable().test(.GET, "/foo?number=true") { res async in
             XCTAssertEqual(res.status, .ok)
             XCTAssertEqual(res.body.string, "42")
-        }.test(.GET, "/foo?number=false") { res in
+        }.test(.GET, "/foo?number=false") { res async in
             XCTAssertEqual(res.status, .ok)
             XCTAssertEqual(res.body.string, "string")
         }
     }
 
-    func testResponseEncodableStatus() throws {
+    func testResponseEncodableStatus() async throws {
         struct User: Content {
             var name: String
         }
-
-        let app = Application(.testing)
-        defer { app.shutdown() }
 
         app.post("users") { req async throws -> Response in
             return try await req.content
@@ -53,7 +57,7 @@ final class AsyncRouteTests: XCTestCase {
                 .encodeResponse(status: .created, for: req)
         }
 
-        try app.testable().test(.POST, "/users", beforeRequest: { req in
+        try await app.testable().test(.POST, "/users", beforeRequest: { req async throws in
             try req.content.encode(["name": "vapor"], as: .json)
         }) { res in
             XCTAssertEqual(res.status, .created)
@@ -64,10 +68,7 @@ final class AsyncRouteTests: XCTestCase {
         }
     }
 
-    func testWebsocketUpgrade() throws {
-        let app = Application(.testing)
-        defer { app.shutdown() }
-
+    func testWebsocketUpgrade() async throws {
         let testMarkerHeaderKey = "TestMarker"
         let testMarkerHeaderValue = "addedInShouldUpgrade"
 
@@ -75,7 +76,7 @@ final class AsyncRouteTests: XCTestCase {
             [testMarkerHeaderKey: testMarkerHeaderValue]
         }, onUpgrade: { _, _ in })
 
-        try app.testable(method: .running(port: 0)).test(.GET, "customshouldupgrade", beforeRequest: { req in
+        try await app.testable(method: .running(port: 0)).test(.GET, "customshouldupgrade", beforeRequest: { req async in
             req.headers.replaceOrAdd(name: HTTPHeaders.Name.secWebSocketVersion, value: "13")
             req.headers.replaceOrAdd(name: HTTPHeaders.Name.secWebSocketKey, value: "zyFJtLIpI2ASsmMHJ4Cf0A==")
             req.headers.replaceOrAdd(name: .connection, value: "Upgrade")
