@@ -56,6 +56,7 @@ final class WebSocketTests: XCTestCase {
             ws.close(promise: nil)
         }
 
+        app.http.server.configuration.port = 0
         app.environment.arguments = ["serve"]
 
         try app.start()
@@ -154,7 +155,7 @@ final class WebSocketTests: XCTestCase {
     }
 }
 
-extension WebSocket: Hashable {
+extension Vapor.WebSocket: Swift.Hashable {
     public static func == (lhs: WebSocket, rhs: WebSocket) -> Bool {
         lhs === rhs
     }
